Initial commit
This is the initial re-implementation of the gsetting module. It works
but there are some limitations.

For example, setting the state to "absent" will reset it to the default
value. Unfortunately idempotence for this state is quite difficult.
This is due to the fact that there does not seem to be a method of
asking if a value is set to use the default.

Unfortunately it is not possible to fix this the usual way, by
comparing previous and new values, since there would be no way to
differentiate between the default and an explicit value that happens
to match the default.

The only condolence is that running reset multiple times is in itself
idempotent. In reality the only thing that is broken is the module
output.

In addition to the common states "present" and "absent", a third one
is included: "get". This is partially to match functionality in the
gconftool2 module, which at this time is the only similar module that
ships with Ansible, making it an example in some respects.