Add backwards compatibility with older dbus
This offers backwards compatibility with systems that ship an older
dbus version; one that does not offer the dbus-run-session utility.

As much as possible of this compatibility is contained in a single
function to simplify future deprecation and removal. This means it is
slightly less efficient, but in reality it is a tiny difference.

Because it uses shell functionality it needs to use run_command with
use_unsafe_shell=True. Tried to find another way, but failed. Will
not pursue this since this is the secondary method.